split things; fix tests
