+++ packages/sure/esm/index.js
@@ -1,4 +1,5 @@
 export * from './core.js';
-export * from './complex.js';
+export * from './object.js';
+export * from './after.js';
 export * from './array.js';
 export * from './primitives.js';
