+++ packages/sure/src/__tests__/array.test.ts
@@ -6,14 +6,12 @@ describe('array', () => {
   it('should return good value', () => {
     const value = someArray([1, 2, 3])
 
-    expect(value).toEqual([true, [1, 2, 3]])
+    expect(value).toStrictEqual([true, [1, 2, 3]])
   })
 
   it('should return evil value', () => {
     const value = someArray([1, 'two', 3])
 
-    expect(value).toEqual([false, ['not number']])
-
-    expect(value).toEqual(evil(['not number']))
+    expect(value).toStrictEqual([false, [undefined, 'not number', undefined]])
   })
 })
